[FEAT] 게시글 삭제(답변 없는 경우에만 가능)
게시글에 답변이 이미 달렸을 때 구체적으로 프론트에서 어떻게 표현할 지는 추후 논의

+++ qna/views.py
@@ -170,6 +170,11 @@ def question_update(request,pk):
 
 def question_delete(request, pk):
     question = get_object_or_404(Question, pk=pk)
+
+    # TODO : 게시글에 댓글이 있는지 확인
+    if len(question.answer_set.all()) > 0:
+        ### 답변이 달려 있어서 삭제 불가능
+        return redirect('qna:question_detail', pk)
     question.delete()
     return redirect('qna:question_list')
